refact(sharepoint-connector): removed unnecessary files and type definitions. moved types out of dedicated types directory
