+++ services/sharepoint-connector/src/msgraph/sharepoint.types.ts
