+++ services/sharepoint-connector/src/unique-api/unique-api.service.ts
@@ -1,9 +1,9 @@
-import { Inject, Injectable, Logger } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
+import {Inject, Injectable, Logger} from '@nestjs/common';
+import {ConfigService} from '@nestjs/config';
 import Bottleneck from 'bottleneck';
-import { GraphQLClient } from 'graphql-request';
-import { Client } from 'undici';
-import { UNIQUE_HTTP_CLIENT } from '../http-client.tokens';
+import {GraphQLClient} from 'graphql-request';
+import {Client} from 'undici';
+import {UNIQUE_HTTP_CLIENT} from '../http-client.tokens';
 import {
   type ContentRegistrationRequest,
   type FileDiffItem,
@@ -11,7 +11,7 @@ import {
   type FileDiffResponse,
   type IngestionApiResponse,
   type IngestionFinalizationRequest,
-} from './types/unique-api.types';
+} from './unique-api.types';
 
 @Injectable()
 export class UniqueApiService {
@@ -23,56 +23,6 @@ export class UniqueApiService {
     @Inject(UNIQUE_HTTP_CLIENT) private readonly httpClient: Client,
   ) {
     this.limiter = new Bottleneck({ maxConcurrent: 1, minTime: 50 });
-  }
-
-  private async makeRateLimitedRequest<T>(requestFn: () => Promise<T>): Promise<T> {
-    return await this.limiter.schedule(async () => await requestFn());
-  }
-
-  private createGraphqlClient(uniqueToken: string): GraphQLClient {
-    const graphqlUrl = this.configService.get<string>('uniqueApi.ingestionGraphQLUrl') ?? '';
-    return new GraphQLClient(graphqlUrl, {
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${uniqueToken}`,
-      },
-    });
-  }
-
-  private getContentUpsertMutation(): string {
-    return `mutation ContentUpsert(
-  $input: ContentCreateInput!
-  $fileUrl: String
-  $chatId: String
-  $scopeId: String
-  $sourceOwnerType: String
-  $sourceName: String
-  $sourceKind: String
-  $storeInternally: Boolean
-) {
-  contentUpsert(
-    input: $input
-    fileUrl: $fileUrl
-    chatId: $chatId
-    scopeId: $scopeId
-    sourceOwnerType: $sourceOwnerType
-    sourceName: $sourceName
-    sourceKind: $sourceKind
-    storeInternally: $storeInternally
-  ) {
-    id
-    key
-    byteSize
-    mimeType
-    ownerType
-    ownerId
-    writeUrl
-    readUrl
-    createdAt
-    internallyStoredAt
-    source { kind name }
-  }
-}`;
   }
 
   public async registerContent(
@@ -118,7 +68,7 @@ export class UniqueApiService {
   ): Promise<FileDiffResponse> {
     return await this.makeRateLimitedRequest(async () => {
       const ingestionUrl = this.configService.get<string>('uniqueApi.ingestionUrl') ?? '';
-      const fileDiffUrl = `${ingestionUrl}/file-diff`;
+      const fileDiffUrl = `${ ingestionUrl }/file-diff`;
       const scopeId = this.configService.get<string>('uniqueApi.scopeId') ?? 'unknown-scope';
       const basePath =
         this.configService.get<string>('uniqueApi.fileDiffBasePath') ??
@@ -143,7 +93,7 @@ export class UniqueApiService {
           path,
           headers: {
             'Content-Type': 'application/json',
-            Authorization: `Bearer ${uniqueToken}`,
+            Authorization: `Bearer ${ uniqueToken }`,
           },
           body: JSON.stringify(diffRequest),
         });
@@ -152,8 +102,8 @@ export class UniqueApiService {
         if (statusCode < 200 || statusCode >= 300) {
           const errorText = await body.text().catch(() => 'No response body');
           throw new Error(
-            `File diff request failed with status ${statusCode}. ` +
-              `URL: ${fileDiffUrl}, Response: ${errorText}`,
+            `File diff request failed with status ${ statusCode }. ` +
+            `URL: ${ fileDiffUrl }, Response: ${ errorText }`,
           );
         }
 
@@ -223,5 +173,55 @@ export class UniqueApiService {
       default:
         return 'SCOPE';
     }
+  }
+
+  private async makeRateLimitedRequest<T>(requestFn: () => Promise<T>): Promise<T> {
+    return await this.limiter.schedule(async () => await requestFn());
+  }
+
+  private createGraphqlClient(uniqueToken: string): GraphQLClient {
+    const graphqlUrl = this.configService.get<string>('uniqueApi.ingestionGraphQLUrl') ?? '';
+    return new GraphQLClient(graphqlUrl, {
+      headers: {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${ uniqueToken }`,
+      },
+    });
+  }
+
+  private getContentUpsertMutation(): string {
+    return `mutation ContentUpsert(
+    $input: ContentCreateInput!
+    $fileUrl: String
+    $chatId: String
+    $scopeId: String
+    $sourceOwnerType: String
+    $sourceName: String
+    $sourceKind: String
+    $storeInternally: Boolean
+    ) {
+    contentUpsert(
+    input: $input
+    fileUrl: $fileUrl
+    chatId: $chatId
+    scopeId: $scopeId
+    sourceOwnerType: $sourceOwnerType
+    sourceName: $sourceName
+    sourceKind: $sourceKind
+    storeInternally: $storeInternally
+    ) {
+    id
+    key
+    byteSize
+    mimeType
+    ownerType
+    ownerId
+    writeUrl
+    readUrl
+    createdAt
+    internallyStoredAt
+    source { kind name }
+    }
+    }`;
   }
 }
